Force backend discovery at instantiation time
Reintroduce the backend discovery at Provider instantation time. Move
remaining items needed for backend handling.
