+++ qiskit/providers/ibmq/providerbackends.py
@@ -16,11 +16,12 @@
 
 import re
 import keyword
-from typing import Iterable
+from typing import Dict, List, Callable, Optional, Any
 from types import SimpleNamespace
 
+from qiskit.providers.providerutils import filter_backends
+
 from .ibmqbackend import IBMQBackend
-from .api.exceptions import RequestsApiError
 
 
 class ProviderBackends(SimpleNamespace):
@@ -32,42 +33,28 @@ class ProviderBackends(SimpleNamespace):
         Args:
             provider (AccountProvider): IBM Q Experience account provider
         """
-        self._provider = provider
-        self._initialized = False
         super().__init__()
+
+        self._provider = provider
+        self._discover_backends()
 
     def _discover_backends(self) -> None:
         """Discovers the remote backends if not already known."""
-        if not self._initialized:
-            try:
-                # Python identifiers can only contain alphanumeric characters
-                # and underscores and cannot start with a digit.
-                pattern = re.compile(r"\W|^(?=\d)", re.ASCII)
-                for backend in self._provider.backends(timeout=3):
-                    backend_name = backend.name()
+        # Python identifiers can only contain alphanumeric characters
+        # and underscores and cannot start with a digit.
+        pattern = re.compile(r"\W|^(?=\d)", re.ASCII)
+        for backend in self._provider._backends.values():
+            backend_name = backend.name()
 
-                    # Make it a valid identifier
-                    if not backend_name.isidentifier():
-                        backend_name = re.sub(pattern, '_', backend_name)
+            # Make it a valid identifier
+            if not backend_name.isidentifier():
+                backend_name = re.sub(pattern, '_', backend_name)
 
-                    # Append _ if is keyword or duplicate
-                    while keyword.iskeyword(backend_name) or backend_name in self.__dict__:
-                        backend_name += '_'
+            # Append _ if is keyword or duplicate
+            while keyword.iskeyword(backend_name) or backend_name in self.__dict__:
+                backend_name += '_'
 
-                    setattr(self, backend_name, backend)
-                self._initialized = True
-            except RequestsApiError:
-                # Ignore any networking errors since this is a convenience
-                # feature meant for interactive sessions.
-                pass
-
-    def __dir__(self) -> Iterable[str]:
-        self._discover_backends()
-        return super().__dir__()
-
-    def __getattr__(self, item: str) -> IBMQBackend:
-        self._discover_backends()
-        return super().__getattribute__(item)
+            setattr(self, backend_name, backend)
 
     def __call__(
             self,
@@ -102,3 +89,22 @@ class ProviderBackends(SimpleNamespace):
             kwargs['backend_name'] = name
 
         return filter_backends(backends, filters=filters, **kwargs)
+
+    @staticmethod
+    def _deprecated_backend_names() -> Dict[str, str]:
+        """Returns deprecated backend names."""
+        return {
+            'ibmqx_qasm_simulator': 'ibmq_qasm_simulator',
+            'ibmqx_hpc_qasm_simulator': 'ibmq_qasm_simulator',
+            'real': 'ibmqx1'
+            }
+
+    @staticmethod
+    def _aliased_backend_names() -> Dict[str, str]:
+        """Returns aliased backend names."""
+        return {
+            'ibmq_5_yorktown': 'ibmqx2',
+            'ibmq_5_tenerife': 'ibmqx4',
+            'ibmq_16_rueschlikon': 'ibmqx5',
+            'ibmq_20_austin': 'QS1_1'
+            }
